+++ examples/with-drizzle/drizzle/schema.ts
@@ -1,0 +1,7 @@
+import { integer, text, sqliteTable } from "drizzle-orm/sqlite-core";
+
+export const Users = sqliteTable("users", {
+  id: integer("id").primaryKey().unique().notNull(),
+  username: text("username").notNull().default(""),
+  password: text("password").notNull().default(""),
+});