+++ packages/start/src/server/pageEvent.ts
@@ -30,6 +30,7 @@ export async function createPageEvent(ctx: FetchEvent) {
     manifest: await clientManifest.json(),
     assets: [
       ...(await clientManifest.inputs[clientManifest.handler]!.assets()),
+      ...(import.meta.env.DEV ? await clientManifest.inputs[import.meta.env.START_APP]!.assets(): []),
       ...(import.meta.env.START_ISLANDS
         ? (await serverManifest.inputs[serverManifest.handler]!.assets()).filter(
             s => (s as any).attrs.rel !== "modulepreload"