+++ packages/start/src/shared/clientOnly.tsx
@@ -3,6 +3,10 @@ import type { Component, ComponentProps, JSX } from "solid-js";
 import { createMemo, createSignal, onMount, sharedConfig, splitProps, untrack } from "solid-js";
 import { isServer } from "solid-js/web";
 
+/**
+ *
+ * Read more: https://docs.solidjs.com/solid-start/reference/client/client-only
+ */
 // not using Suspense
 export default function clientOnly<T extends Component<any>>(
   fn: () => Promise<{